♻️ Client.installLoadingGraph => add customInstallers handling + aliases definition
